+++ ako-operator/controllers/gatewayclass.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2023 VMware, Inc.
+Copyright © 2025 Broadcom Inc. and/or its subsidiaries. All Rights Reserved.
 All Rights Reserved.
 
 Licensed under the Apache License, Version 2.0 (the "License");