+++ internal/cache/controller_obj_cache.go
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019-2020 VMware, Inc.
+ * Copyright © 2025 Broadcom Inc. and/or its subsidiaries. All Rights Reserved.
  * All Rights Reserved.
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
@@ -102,7 +102,7 @@ func (c *AviObjCache) AviRefreshObjectCache(client []*clients.AviClient, cloud s
 		c.PopulateVsVipDataToCache(client[7], cloud)
 	}()
 	c.PopulatePkiProfilesToCache(client[0])
-	c.PopulateAppPersistenceProfileToCache(client[9], cloud)
+	c.PopulateAppPersistenceProfileToCache(client[0])
 	c.PopulatePoolsToCache(client[1], cloud)
 	c.PopulatePgDataToCache(client[2], cloud)
 	c.PopulateStringGroupDataToCache(client[8], cloud)
@@ -1224,6 +1224,56 @@ func (c *AviObjCache) AviPopulateOnePKICache(client *clients.AviClient,
 	return nil
 }
 
+func (c *AviObjCache) AviPopulateOnePersistenceProfileCache(client *clients.AviClient, objName string) error {
+	var uri string
+	uri = "/api/applicationpersistenceprofile?name=" + objName + "&include_name=true"
+	result, err := lib.AviGetCollectionRaw(client, uri)
+	if err != nil {
+		utils.AviLog.Warnf("Get uri %v returned err for persistenceprofile %v", uri, err)
+		return err
+	}
+	elems := make([]json.RawMessage, result.Count)
+	err = json.Unmarshal(result.Results, &elems)
+	if err != nil {
+		utils.AviLog.Warnf("Failed to unmarshal persistenceprofile data, err: %v", err)
+		return err
+	}
+	for i := 0; i < len(elems); i++ {
+		persistenceProfileKey := models.ApplicationPersistenceProfile{}
+		err = json.Unmarshal(elems[i], &persistenceProfileKey)
+		if err != nil {
+			utils.AviLog.Warnf("Failed to unmarshal persistenceprofile data, err: %v", err)
+			continue
+		}
+		if persistenceProfileKey.Name == nil || persistenceProfileKey.UUID == nil {
+			utils.AviLog.Warnf("Incomplete persistenceprofile data unmarshalled, %s", utils.Stringify(persistenceProfileKey))
+			continue
+		}
+		//Only cache a Persistence Profile key that belongs to this AKO.
+		if !strings.HasPrefix(*persistenceProfileKey.Name, lib.GetNamePrefix()) {
+			continue
+		}
+		emptyIngestionMarkers := utils.AviObjectMarkers{}
+		tenant := getTenantFromTenantRef(*persistenceProfileKey.TenantRef)
+		chksum := lib.PersistenceProfileChecksum(*persistenceProfileKey.Name, *persistenceProfileKey.PersistenceType, emptyIngestionMarkers, persistenceProfileKey.Markers, true)
+		if persistenceProfileKey.HTTPCookiePersistenceProfile != nil {
+			chksum += lib.HTTPCookiePersistenceProfileChecksum(*persistenceProfileKey.HTTPCookiePersistenceProfile.CookieName, persistenceProfileKey.HTTPCookiePersistenceProfile.Timeout, persistenceProfileKey.HTTPCookiePersistenceProfile.IsPersistentCookie)
+		}
+		cacheObj := AviPersistenceProfileCache{
+			Name:             *persistenceProfileKey.Name,
+			Tenant:           tenant,
+			Uuid:             *persistenceProfileKey.UUID,
+			CloudConfigCksum: chksum,
+			Type:             *persistenceProfileKey.PersistenceType,
+			LastModified:     *persistenceProfileKey.LastModified,
+		}
+		k := NamespaceName{Namespace: tenant, Name: *persistenceProfileKey.Name}
+		c.AppPersProfileCache.AviCacheAdd(k, &cacheObj)
+		utils.AviLog.Debugf("Adding persistenceprofile to Cache during refresh %s", k)
+	}
+	return nil
+}
+
 func (c *AviObjCache) AviPopulateOnePoolCache(client *clients.AviClient,
 	cloud string, objName string) error {
 	var uri string
@@ -2164,7 +2214,7 @@ func (c *AviObjCache) AviPopulateOneStringGroupCache(client *clients.AviClient,
 	return nil
 }
 
-func (c *AviObjCache) AviPopulateAllAppPersistenceProfiles(client *clients.AviClient, cloud string, appPersProfileData *[]AviPersistenceProfileCache, nextPage ...NextPage) (*[]AviPersistenceProfileCache, int, error) {
+func (c *AviObjCache) AviPopulateAllAppPersistenceProfiles(client *clients.AviClient, appPersProfileData *[]AviPersistenceProfileCache, nextPage ...NextPage) (*[]AviPersistenceProfileCache, int, error) {
 	var uri string
 	if len(nextPage) == 1 {
 		uri = nextPage[0].NextURI
@@ -2217,7 +2267,7 @@ func (c *AviObjCache) AviPopulateAllAppPersistenceProfiles(client *clients.AviCl
 		if len(next_uri) > 1 {
 			overrideUri := "/api/applicationpersistenceprofile" + next_uri[1]
 			nextPage := NextPage{NextURI: overrideUri}
-			_, _, err := c.AviPopulateAllAppPersistenceProfiles(client, cloud, appPersProfileData, nextPage)
+			_, _, err := c.AviPopulateAllAppPersistenceProfiles(client, appPersProfileData, nextPage)
 			if err != nil {
 				return nil, 0, err
 			}
@@ -2226,13 +2276,13 @@ func (c *AviObjCache) AviPopulateAllAppPersistenceProfiles(client *clients.AviCl
 	return appPersProfileData, result.Count, nil
 }
 
-func (c *AviObjCache) PopulateAppPersistenceProfileToCache(client *clients.AviClient, cloud string) {
+func (c *AviObjCache) PopulateAppPersistenceProfileToCache(client *clients.AviClient) {
 	var appPersProfileData []AviPersistenceProfileCache
 	setDefaultTenant := session.SetTenant(lib.GetTenant())
 	setTenant := session.SetTenant("*")
 	setTenant(client.AviSession)
 	defer setDefaultTenant(client.AviSession)
-	c.AviPopulateAllAppPersistenceProfiles(client, cloud, &appPersProfileData)
+	c.AviPopulateAllAppPersistenceProfiles(client, &appPersProfileData)
 
 	persistenceCacheData := c.AppPersProfileCache.ShallowCopy()
 	for i, persistenceCacheObj := range appPersProfileData {
@@ -3955,7 +4005,7 @@ func checkIPAMForUsableNetworkLabels(client *clients.AviClient, ipamRefUri *stri
 	}
 
 	// 4. Empty VipNetworkList
-	if utils.IsWCP() && markerNetworkFound == "" {
+	if (utils.IsWCP() && markerNetworkFound == "") || lib.GetVPCMode() {
 		utils.SetVipNetworkList([]akov1beta1.AviInfraSettingVipNetwork{})
 		return true, nil
 	}